Removed one instance of drwl_rect function.
I believe the way this function was written is flawed. I am working on
seperating it's logic and keeping the parameters manageable. This means
I will eventually seperate the function drawbar(Monitor *) into a
multitude of functions each part handling a seperate part of the
statusbar. This is far more readable in my opinion. Is this code review?
Who knows!