Just a temporary fix.
The statusbar's background would not be colored. Thus leaving a strange
black background that didn't fit aesthetically when a window was open.
This temporarily (potentially permanent) fix addresses this issue
through writing the statusbar last. This is to let the top bar be drawn
in without conflicting with the statusbar. The only issue that may arise
is the title of the window may encroach on the statusbar.